+++ redis-single/CHANGELOG.md
@@ -3,3 +3,5 @@
 * Initial redis-single image
 * Fix consul config for multiple services
 * Remove invalid log level for redis_exporter
+* Add redis log to syslog-ng
+* Set option for redis_exporter if authpass set